+++ src/Traits/Classic/HasVerifiedAtHelpers.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Cog\Flag\Traits\Classic;
 
-use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Date;
 
 trait HasVerifiedAtHelpers
 {
@@ -29,7 +29,7 @@ trait HasVerifiedAtHelpers
      */
     public function setVerifiedFlag()
     {
-        $this->setAttribute('verified_at', Carbon::now());
+        $this->setAttribute('verified_at', Date::now());
 
         return $this;
     }