+++ src/Traits/Inverse/HasClosedAtHelpers.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Cog\Flag\Traits\Inverse;
 
-use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Date;
 
 trait HasClosedAtHelpers
 {
@@ -29,7 +29,7 @@ trait HasClosedAtHelpers
      */
     public function setClosedFlag()
     {
-        $this->setAttribute('closed_at', Carbon::now());
+        $this->setAttribute('closed_at', Date::now());
 
         return $this;
     }