+++ tests/Unit/Scopes/Inverse/EndedAtScopeTest.php
@@ -16,7 +16,7 @@ namespace Cog\Tests\Flag\Unit\Scopes\Inverse;
 use Cog\Tests\Flag\Stubs\Models\Inverse\EntityWithEndedAt;
 use Cog\Tests\Flag\Stubs\Models\Inverse\EntityWithEndedAtUnapplied;
 use Cog\Tests\Flag\TestCase;
-use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Date;
 
 final class EndedAtScopeTest extends TestCase
 {
@@ -24,7 +24,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_get_only_not_ended(): void
     {
         factory(EntityWithEndedAt::class, 2)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
         factory(EntityWithEndedAt::class, 3)->create([
             'ended_at' => null,
@@ -39,7 +39,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_get_without_ended(): void
     {
         factory(EntityWithEndedAt::class, 2)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
         factory(EntityWithEndedAt::class, 3)->create([
             'ended_at' => null,
@@ -54,7 +54,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_get_with_ended(): void
     {
         factory(EntityWithEndedAt::class, 2)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
         factory(EntityWithEndedAt::class, 3)->create([
             'ended_at' => null,
@@ -69,7 +69,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_get_only_ended(): void
     {
         factory(EntityWithEndedAt::class, 2)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
         factory(EntityWithEndedAt::class, 3)->create([
             'ended_at' => null,
@@ -84,7 +84,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_unend_model(): void
     {
         $model = factory(EntityWithEndedAt::class)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
 
         EntityWithEndedAt::where('id', $model->id)->unend();
@@ -112,7 +112,7 @@ final class EndedAtScopeTest extends TestCase
     public function it_can_skip_apply(): void
     {
         factory(EntityWithEndedAt::class, 3)->create([
-            'ended_at' => Carbon::now()->subDay(),
+            'ended_at' => Date::now()->subDay(),
         ]);
         factory(EntityWithEndedAt::class, 2)->create([
             'ended_at' => null,