+++ CHANGELOG.md
@@ -12,12 +12,12 @@ All notable changes to `laravel-eloquent-flag` will be documented in this file.
 ### Changed
 
 - All methods are strict typed now
-- `Carbon\Carbon` replaced with `Illuminate\Support\Carbon`
+- `Carbon\Carbon` replaced with `Illuminate\Support\Facades\Date`
 
 ### Removed
 
 - Dropped PHP 5.6, 7.0 support
-- Dropped Laravel 5.2, 5.3, 5.4, 5.5, 5.6 support
+- Dropped Laravel 5.2, 5.3, 5.4, 5.5, 5.6, 5.7 support
 
 ## [4.0.0] - 2018-09-09